\docType{package}
\name{ganalytics}
\alias{ganalytics}
\alias{ganalytics-package}
\title{gprediction}
\description{
  Google Prediction API client for R
}
\details{
  A Google Prediction API client for R supporting OAuth2.0
}
\references{
  Google Predication API developer reference
  \url{https://developers.google.com/prediction/docs/getting-started}
}
\keyword{package}

